feat: Shifts Module
